+++ tests/unit/butterfree/extract/readers/test_file_reader.py
@@ -2,9 +2,7 @@ import pytest
 from pyspark.sql.types import DoubleType, LongType, StringType, StructField, StructType
 
 from butterfree.clients import SparkClient
-from butterfree.dataframe_service import IncrementalStrategy
 from butterfree.extract.readers import FileReader
-from butterfree.metadata.reader_metadata import FileReaderMetadata
 
 
 class TestFileReader:
@@ -130,37 +128,3 @@ class TestFileReader:
         assert df.columns == ["A", "B", "C"]
         for value in range(3):
             assert df.first()[value] != ["A", "B", "C"][value]
-
-    def test_build_metadata(self):
-        # given
-        file_reader = FileReader(
-            id="file_reader",
-            path="path",
-            format="format",
-        )
-
-        # when
-        metadata = file_reader.build_metadata()
-
-        # then
-        assert isinstance(metadata, FileReaderMetadata)
-        assert metadata.path == file_reader.path
-        assert metadata.format == file_reader.format
-        assert not metadata.incremental_strategy
-
-    def test_build_metadata_with_incremental_strategy(self):
-        # given
-        file_reader = FileReader(
-            id="file_reader",
-            path="path",
-            format="format",
-        ).with_incremental_strategy(IncrementalStrategy(column="timestamp"))
-
-        # when
-        metadata = file_reader.build_metadata()
-
-        # then
-        assert isinstance(metadata, FileReaderMetadata)
-        assert metadata.path == file_reader.path
-        assert metadata.format == file_reader.format
-        assert metadata.incremental_strategy
