+++ tests/unit/butterfree/extract/readers/test_table_reader.py
@@ -1,8 +1,6 @@
 import pytest
 
-from butterfree.dataframe_service import IncrementalStrategy
 from butterfree.extract.readers import TableReader
-from butterfree.metadata.reader_metadata import TableReaderMetadata
 
 
 class TestTableReader:
@@ -34,37 +32,3 @@ class TestTableReader:
         # assert
         spark_client.read_table.assert_called_once_with(table, database)
         assert target_df.collect() == output_df.collect()
-
-    def test_build_metadata(self):
-        # given
-        table_reader = TableReader(
-            id="table_reader",
-            database="db",
-            table="table",
-        )
-
-        # when
-        metadata = table_reader.build_metadata()
-
-        # then
-        assert isinstance(metadata, TableReaderMetadata)
-        assert metadata.database == table_reader.database
-        assert metadata.table == table_reader.table
-        assert not metadata.incremental_strategy
-
-    def test_build_metadata_with_incremental_strategy(self):
-        # given
-        table_reader = TableReader(
-            id="table_reader",
-            database="db",
-            table="table",
-        ).with_incremental_strategy(IncrementalStrategy(column="timestamp"))
-
-        # when
-        metadata = table_reader.build_metadata()
-
-        # then
-        assert isinstance(metadata, TableReaderMetadata)
-        assert metadata.database == table_reader.database
-        assert metadata.table == table_reader.table
-        assert metadata.incremental_strategy
