+++ tests/unit/butterfree/load/writers/test_online_feature_store_writer.py
@@ -8,7 +8,6 @@ from butterfree.clients import SparkClient
 from butterfree.configs.db import CassandraConfig, KafkaConfig
 from butterfree.load.processing import json_transform
 from butterfree.load.writers import OnlineFeatureStoreWriter
-from butterfree.metadata.writer_metadata import WriterMetadata
 from butterfree.testing.dataframe import assert_dataframe_equality
 
 
@@ -302,17 +301,3 @@ class TestOnlineFeatureStoreWriter:
                 topic="custom_topic"
             ).items()
         )
-
-    def test_build_metadata(self):
-        # given
-        writer = OnlineFeatureStoreWriter()
-
-        # when
-        metadata = writer.build_metadata()
-
-        # then
-        assert isinstance(metadata, WriterMetadata)
-        assert metadata.type == "OnlineFeatureStoreWriter"
-        assert metadata.interval_mode == writer.interval_mode
-        assert metadata.write_to_entity == writer.write_to_entity
-        assert metadata.db_config == writer.db_config.__class__.__name__
