+++ tests/unit/butterfree/transform/features/test_key_feature.py
@@ -1,7 +1,6 @@
 from unittest.mock import Mock
 
 from butterfree.constants import DataType
-from butterfree.metadata.feature_metadata import FeatureMetadata
 from butterfree.transform.features import KeyFeature
 
 
@@ -30,21 +29,3 @@ class TestKeyFeature:
         assert test_key.from_column == "origin"
         assert test_key.description == "unit test"
         assert test_key.transformation
-
-    def test_build_metadata(self):
-        # arrange
-        key_feature = KeyFeature(
-            name="my_key",
-            description="a key feature",
-            dtype=DataType.INTEGER,
-        )
-
-        # act
-        metadata = key_feature.build_metadata()
-
-        # assert
-        assert isinstance(metadata, FeatureMetadata)
-        assert metadata.name == "my_key"
-        assert metadata.data_type == "INTEGER"
-        assert metadata.primary_key is True
-        assert metadata.description == "a key feature"
